+++ packages/nuxt-json/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @contentrain/nuxt-json
 
+## 1.0.4
+
+- Clean unnececary logs
+
+## 1.0.3
+
+- Storage to server assets fix 2
+
+## 1.0.2
+
+### Patch Changes
+
+- Storage to server assets fix
+
+## 1.0.1
+
+### Patch Changes
+
+- Storage to server assets
+
 ## 1.0.0
 
 ### Major Changes